+++ common/components/screen-editor/src/commonMain/kotlin/io/stardewvalleydesigner/component/editor/modules/toolkit/ToolType.kt
@@ -26,4 +26,5 @@ enum class ToolType {
     Eraser,
     Select,
     EyeDropper,
+    Hand,
 }